python/✅day18/requirements.txt
heroes
colorgram.py